cloud_functions/heartbeat_handler/requirements.txt
google-cloud-firestore
geohash2